requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-templated-mail==1.1.1
django-timezone-field==5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
exceptiongroup==1.1.2
factory-boy==3.2.1
Faker==19.1.0
flake8==6.0.0
flake8-broken-line==1.0.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.3
flake8-return==1.2.0
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
kombu==5.3.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
pep8-naming==0.13.3
pluggy==1.2.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-crontab==3.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
